tools: tools.py deleted
    - The aim is to boil the 'core' source code down to its bare
      elements. To separate testing code from change detection code,
      tools.py has been eliminated. To facilitate this change the
      following functions have been moved out of tools.py:

          - gendata(), plotprob() and plotbound() moved into tests.py
          - filterdata() and segmentdata() moved into tests.py
